- fix issue https://github.com/bigbluebutton/bigbluebutton/issues/11358 by having the parent meeting    tell the breakout rooms the remaining time. Currently, each room has its own countdown timer which    could introduce drifting resulting in breakout rooms ending at different times.
