+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core2/message/handlers/SendBreakoutTimeRemainingMsgHdlr.scala
@@ -1,29 +1,36 @@
 package org.bigbluebutton.core2.message.handlers
 
 import org.bigbluebutton.common2.msgs._
-import org.bigbluebutton.core.api.SendTimeRemainingAuditInternalMsg
+import org.bigbluebutton.core.api.{ EndBreakoutRoomInternalMsg, SendBreakoutTimeRemainingInternalMsg, SendTimeRemainingAuditInternalMsg }
+import org.bigbluebutton.core.bus.BigBlueButtonEvent
 import org.bigbluebutton.core.domain.MeetingState2x
-import org.bigbluebutton.core.running.{ LiveMeeting, OutMsgRouter }
+import org.bigbluebutton.core.running.{ LiveMeeting, MeetingActor, OutMsgRouter }
 import org.bigbluebutton.core.util.TimeUtil
 
 trait SendBreakoutTimeRemainingMsgHdlr {
+  this: MeetingActor =>
+
   val liveMeeting: LiveMeeting
   val outGW: OutMsgRouter
 
   def handleSendBreakoutTimeRemainingMsg(msg: SendTimeRemainingAuditInternalMsg, state: MeetingState2x): MeetingState2x = {
-
     for {
       model <- state.breakout
       startedOn <- model.startedOn
     } yield {
+      val endMeetingTime = TimeUtil.millisToSeconds(startedOn) + TimeUtil.minutesToSeconds(model.durationInMinutes)
+      val timeRemaining = endMeetingTime - TimeUtil.millisToSeconds(System.currentTimeMillis())
+
       if (!liveMeeting.props.meetingProp.isBreakout) {
-
-        val endMeetingTime = TimeUtil.millisToSeconds(startedOn) + TimeUtil.minutesToSeconds(model.durationInMinutes)
-        val timeRemaining = endMeetingTime - TimeUtil.millisToSeconds(System.currentTimeMillis())
-
+        // Notify parent meeting users of breakout rooms time remaining
         val event = buildBreakoutRoomsTimeRemainingUpdateEvtMsg(liveMeeting.props.meetingProp.intId, timeRemaining.toInt)
-
         outGW.send(event)
+      }
+
+      // Tell all breakout rooms of time remaining so they can notify their users.
+      // This syncs all rooms about time remaining.
+      model.rooms.values.foreach { room =>
+        eventBus.publish(BigBlueButtonEvent(room.id, SendBreakoutTimeRemainingInternalMsg(props.breakoutProps.parentId, timeRemaining.toInt)))
       }
     }
